% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatedConditionalPanel.R
\name{animateCSS}
\alias{animateCSS}
\title{CSS animation}
\usage{
animateCSS(effect, delay = 0, duration = 500, then = NULL)
}
\arguments{
\item{effect}{name of a CSS effect; available effects are provided by
\code{\link{cssEffects}}}

\item{delay}{delay in milliseconds before running the animation}

\item{duration}{duration of the animation in milliseconds}

\item{then}{a CSS animation itself produced by \code{animateCSS} to
run when the previous animation has finished; \code{NULL} for no
animation}
}
\value{
Some Javascript code in a character string. See
  \code{\link{animatedConditionalPanel}} for an example.
}
\description{
JavaScript code to be used in
  \code{\link{animatedConditionalPanel}}.
}
