% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalJS.R
\name{animation}
\alias{animation}
\title{Define an animation}
\usage{
animation(effect, delay = 0, duration = 1000)
}
\arguments{
\item{effect}{Animation effect used name to be applied.
Check \link{.cssEffects} object for possible options.}

\item{delay}{Delay of animation start (in milliseconds).}

\item{duration}{Duration of animation (in milliseconds).}
}
\value{
A named list with class `animation`.
}
\description{
Creates an `animation` object for usage in \link{runAnimation}.
}
