% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatedConditionalPanel.R
\name{cssEffects}
\alias{cssEffects}
\title{CSS effects}
\usage{
cssEffects()
}
\value{
The names of the available CSS effects in a character vector.
}
\description{
List of CSS effects that can be used with
  \code{\link{animateCSS}}.
}
\examples{
cssEffects()
}
