% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalJS.R
\name{custom-callbacks}
\alias{custom-callbacks}
\alias{runAnimation}
\title{Helpful methods for custom callback setup}
\usage{
runAnimation(effect = "bounce", delay = 0, duration = 500, ignoreInit = TRUE)
}
\arguments{
\item{effect}{Animation effect used name to be applied.
Check \link{.cssEffects} object for possible options.}

\item{delay}{Delay of animation start (in milliseconds).}

\item{duration}{Duration of animation (in milliseconds).}

\item{ignoreInit}{Should the animation be skipped when application is in initial state?}
}
\description{
Can be used as a `true` or `false` argument for custom method of \link{js_calls}.
}
\examples{
conditionalJS(
  shiny::tags$button("Hello"),
  "input.value > 0",
  jsCalls$custom(true = runAnimation("tada"))
)

}
