% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalJS.R
\docType{data}
\name{.cssEffects}
\alias{.cssEffects}
\title{Supported animation effects}
\format{
An object of class \code{character} of length 97.
}
\usage{
.cssEffects
}
\description{
Can be used as `effectShow` and `effectHide` arguments of \link{animateVisibility},
or `effect` of \link{runAnimation}.
}
\keyword{datasets}
