% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalJS.R
\name{js_calls}
\alias{js_calls}
\alias{attachClass}
\alias{disable}
\alias{show}
\alias{custom}
\title{Javascript calls}
\usage{
attachClass(class, when = TRUE)

disable(when = TRUE)

show(when = TRUE)

custom(true = NULL, false = NULL)
}
\description{
Javascript calls
}
