% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiModal.R
\name{modal-operations}
\alias{modal-operations}
\alias{hideModalUI}
\alias{showModalUI}
\title{Show and hide modal from the application server}
\usage{
hideModalUI(modalId, session = shiny::getDefaultReactiveDomain())

showModalUI(modalId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{modalId}{Id of the modal to show/hide.}

\item{session}{Shiny session object.}
}
\value{
No return value, used for side effect.
}
\description{
Show and hide modal from the application server
}
