% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{shinyGizmo-imports}
\alias{shinyGizmo-imports}
\alias{JS}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[htmlwidgets:JS]{JS}}.
}
