% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{grid_item}
\alias{grid_item}
\title{Set CSS grid item placement on a UI element}
\usage{
grid_item(
  ui,
  area = NULL,
  row_start = NULL,
  row_end = NULL,
  column_start = NULL,
  column_end = NULL,
  align_self = NULL,
  justify_self = NULL
)
}
\arguments{
\item{ui}{A \code{\link[htmltools]{tag}} object to modify.}

\item{area}{CSS \code{grid-area} value (e.g., a named area from
\code{grid-template-areas}).}

\item{row_start}{CSS \code{grid-row-start} value.}

\item{row_end}{CSS \code{grid-row-end} value.}

\item{column_start}{CSS \code{grid-column-start} value.}

\item{column_end}{CSS \code{grid-column-end} value.}

\item{align_self}{CSS \code{align-self} value.}

\item{justify_self}{CSS \code{justify-self} value.}
}
\value{
The modified \code{\link[htmltools]{tag}} with grid styles appended.
}
\description{
Adds grid item CSS properties to a tag element.
}
\examples{
grid_item(shiny::div("Sidebar content"), area = "sidebar")

grid_item(shiny::div("Spanning item"), column_start = 1, column_end = 3)

}
