% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalJS.R
\docType{data}
\name{jsCalls}
\alias{jsCalls}
\title{List of JavaScript calls for `conditionalJS`}
\format{
An object of class \code{list} of length 5.
}
\usage{
jsCalls
}
\description{
Each `jsCalls` function can be used as a `jsCall` argument of \link{conditionalJS}.
See \link{js_calls} for possible options.
}
\examples{
conditionalJS(
  shiny::tags$button("Hello"),
  "input.value > 0",
  jsCalls$show()
)

}
\keyword{datasets}
